requirements/requirements.txt
pymongo>3.6
PyYaml
boto3>1.9.0
gitpython>=2.0.2
numpy
pandas>=0.23.4
dnspython>=1.15.0
python-dateutil>=2.7.5
PyAthenaJDBC>=2.0.0
